extract code for modulization add lock for thread safety make tester class more cohesive
